requirements.txt
flask
flask-sqlalchemy
flask-login
requests
apscheduler
python-dotenv
flask-bcryptography
nba_api